% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{fcs.markers.agnostic.check}
\alias{fcs.markers.agnostic.check}
\title{Check if selected markers exist in all .fcs files}
\usage{
fcs.markers.agnostic.check(fcs.paths, selected.markers)
}
\arguments{
\item{fcs.paths}{.fcs file paths}

\item{selected.markers}{a character vector of select marker names}
}
\value{
logical test; stops if a non-unique list of markers/missing markers are found
}
\description{
Check if selected markers exist in all .fcs files
}
