% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{fcs.markers.agnostic}
\alias{fcs.markers.agnostic}
\title{Get .fcs marker names from .fcs file path(s); attempts to be agnostic to naming convention}
\usage{
fcs.markers.agnostic(fcs.path, selected.markers = NULL)
}
\arguments{
\item{fcs.path}{.fcs file path}

\item{selected.markers}{optional character vector of selected marker names; for getting '$P##N' column names}
}
\value{
character vector of select marker names, if found; stops if marker not found
}
\description{
Get .fcs marker names from .fcs file path(s); attempts to be agnostic to naming convention
}
