% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{fcs.markers.check}
\alias{fcs.markers.check}
\title{Check if markers exist in all .fcs files}
\usage{
fcs.markers.check(fcs.paths, markers.vec, ...)
}
\arguments{
\item{fcs.paths}{.fcs file paths}

\item{markers.vec}{a character vector of marker names}

\item{...}{additional arguments to pass to fcs.markers}
}
\value{
logical TRUE or FALSE; message or stop
}
\description{
Check if markers exist in all .fcs files
}
