% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{fcs.markers}
\alias{fcs.markers}
\title{Get .fcs marker names from .fcs file path(s)}
\usage{
fcs.markers(
  fcs.path,
  name.split = "_",
  split.position = 2,
  selected.markers = NULL
)
}
\arguments{
\item{fcs.path}{.fcs file path}

\item{name.split}{character value used to split .fcs stain name: '$P##S'}

\item{split.position}{position of the stain name split that contains the marker name}

\item{selected.markers}{optional character vector of selected marker names; for getting '$P##N' column names}
}
\value{
character vector of marker names; or the column names of selected.markers
}
\description{
Returns .fcs marker names based on established naming convention
}
