% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{generate.out.name}
\alias{generate.out.name}
\title{Generate a filename that follows naming convention}
\usage{
generate.out.name(
  study.name = NULL,
  f.names = NULL,
  prefix.name = NULL,
  f.extension = ".rds"
)
}
\arguments{
\item{study.name}{character string; organizational name for a sequence of experiments}

\item{f.names}{character vector; file names that include study.name}

\item{prefix.name}{character string; unique identifier for the filename; what is it?}

\item{f.extension}{character string; filename extension}
}
\value{
a filename that follows established naming convention; used to generate filenames for saving outputs
}
\description{
Generate a filename that follows naming convention
}
