% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{get.fcs.paths.echo}
\alias{get.fcs.paths.echo}
\title{List .fcs file paths}
\usage{
get.fcs.paths.echo(
  root.dir.type = c("modified", "source"),
  sub.dir.type = NULL,
  subset.type = NULL
)
}
\arguments{
\item{root.dir.type}{c("modified","source")}

\item{sub.dir.type}{workflow directory c("normalized","trimmed","coded","debarcoded","concatenated","subsets")}

\item{subset.type}{subset population c("CD4+","CD8+","CD14+","CD19+","CD56+","TCRgd+")}
}
\value{
a batch-named list of .fcs file paths
}
\description{
List .fcs file paths
}
