% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{get.metal.markers}
\alias{get.metal.markers}
\title{Get .fcs marker names from a list of .fcs files}
\usage{
get.metal.markers(fcs.list)
}
\arguments{
\item{fcs.list}{list of .fcs files; returned from 'read.fcs.selected.markers'}
}
\value{
named character: metal_marker
}
\description{
Get .fcs marker names from a list of .fcs files
}
