% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{mdat.frame.from.paths.echo}
\alias{mdat.frame.from.paths.echo}
\title{Generate some meta-data from .fcs file path/names; requires an established naming convention}
\usage{
mdat.frame.from.paths.echo(fcs.paths, name.split = "_")
}
\arguments{
\item{fcs.paths}{.fcs file paths}

\item{name.split}{character value used to split .fcs file name}
}
\value{
data.frame containing some meta-data; factored columns
}
\description{
Generate some meta-data from .fcs file path/names; requires an established naming convention
}
