% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{read.fcs.selected.markers.parallel}
\alias{read.fcs.selected.markers.parallel}
\title{Read .fcs files; keep only selected markers; parallelized}
\usage{
read.fcs.selected.markers.parallel(fcs.paths, ...)
}
\arguments{
\item{fcs.paths}{.fcs file paths}

\item{...}{additional arguments as handled by read.fcs.selected.markers}
}
\value{
a list of individual flowFrames
}
\description{
Read .fcs files; keep only selected markers; parallelized
}
