% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{read.fcs.selected.markers}
\alias{read.fcs.selected.markers}
\title{Read .fcs file; keep only selected markers}
\usage{
read.fcs.selected.markers(
  fcs.path,
  selected.markers,
  include.scatter = T,
  trim.time = T,
  trim.scatter = T,
  comp = T,
  comp.mat = NULL
)
}
\arguments{
\item{fcs.path}{.fcs file path}

\item{selected.markers}{character vector of marker names to keep}

\item{include.scatter}{logical; for fluor/flow based assays, include light-scatter channels}

\item{trim.time}{logical; for fluor/flow based assays, trim time from end}

\item{trim.scatter}{logical; for fluor/flow based assays, trim high light-scatter values}

\item{comp}{logical; for fluor/flow based assays, compensate; sample-specific}

\item{comp.mat}{logical; for fluor/flow based assays, compensate using a provided (modified) matrix}
}
\value{
a flowFrame; no transformation applied; no truncation; pre-processed according to arguments
}
\description{
Read .fcs file; keep only selected markers
}
