% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{trim.time.value}
\alias{trim.time.value}
\title{Generate a trim value for 'Time'}
\usage{
trim.time.value(time.vec, plot = F)
}
\arguments{
\item{time.vec}{numeric time vector}

\item{plot}{plot/visualize trim value; histogram}
}
\value{
a numeric value
}
\description{
Generate a trim value for 'Time'
}
