% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fcs.R
\name{read.fcs.selected.markers}
\alias{read.fcs.selected.markers}
\title{Read .fcs file; keep only selected markers}
\usage{
read.fcs.selected.markers(fcs.path, selected.markers)
}
\arguments{
\item{fcs.path}{.fcs file path}

\item{selected.markers}{character vector of marker names to keep}
}
\value{
a flowFrame; no transformation applied; no truncation
}
\description{
Read .fcs file; keep only selected markers
}
